requirements.txt
asgiref==3.9.1
black==24.8.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.2
colorama==0.4.6
cryptography==45.0.6
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.2.2
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
flake8==7.1.1
google-auth==2.40.3
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
isort==5.13.2
mccabe==0.7.0
mypy_extensions==1.1.0
mysqlclient==2.2.4
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.8
pluggy==1.6.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.10.1
pytest==8.3.2
pytest-django==4.8.0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
whitenoise==6.11.0
